Un-nest local schema from plum namespace, refs #339
