+++ app/models/schema/ephemera.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 ##
-# Defining the attributes for Ephemera Resources within Plum
+# Defining the attributes for Ephemera Resources
 module Schema
   module Ephemera
     extend ActiveSupport::Concern
 
     def self.attributes
-      SCHEMA.attributes + RDFS.attributes + Plum::Local.attributes + NFO.attributes + OpaqueMods.attributes + DublinCore.attributes + BIBFRAME.attributes
+      SCHEMA.attributes + RDFS.attributes + Local.attributes + NFO.attributes + OpaqueMods.attributes + DublinCore.attributes + BIBFRAME.attributes
     end
 
     included do
